requirements.txt
boto3==1.24.93
botocore==1.27.93
certifi==2022.12.7
charset-normalizer==2.1.1
docutils==0.19
idna==3.4
jmespath==1.0.1
lxml==4.9.2
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
python-dateutil==2.8.2
python-pptx==0.6.21
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
